auth-service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
python-decouple==3.8
drf-yasg==1.21.7
django-ratelimit==4.1.0